opq_enclosing

Find all features which enclose a given point, and optionally match specific 'key'-'value' pairs. This function is not intended to be combined with add_osm_feature, rather is only to be used in the sequence opq_enclosing -> opq_string -> osmdata_xml (or other extraction function). See examples for how to use.

opq_enclosing(
  lon,
  lat,
  key = NULL,
  value = NULL,
  enclosing = "relation",
  timeout = 25
)

Arguments

lon
Longitude of desired point

lat
Latitude of desired point

key
(Optional) OSM key of enclosing data

value
(Optional) OSM value matching 'key' of enclosing data

enclosing
Either 'relation' or 'way' for whether to return enclosing objects of those respective types (where generally 'relation' will correspond to multipolygon objects, and 'way' to polygon objects).

timeout
It may be necessary to increase this value for large queries, because the server may time out before all data are delivered.

Examples

if (FALSE) {
# Get water body surrounding a particular point:
lat <- 54.33601
lon <- -3.07677
key <- "natural"
value <- "water"
x <- opq_enclosing (lon, lat, key, value) %>%
    opq_string () %>%
    osmdata_sf ()
}