osm_elevation

Add elevation data to a previously-extracted OSM data set, using a pre-downloaded global elevation file from http://srtm.csi.cgiar.org/srtmdata. Currently only works for SC-class objects returned from osmdata_sc.

osm_elevation(dat, elev_file)

Arguments

dat
An SC object produced by osmdata_sc.

elev_file
A vector of one or more character strings specifying paths to .tif files containing global elevation data.

Value

A modified version of the input dat with an additional z_ column appended to the vertices.