Extract all osm_line objects

If id is of a point object, osm_lines will return all lines containing that point. If id is of a line or polygon object, osm_lines will return all lines which intersect the given line or polygon.

osm_lines(dat, id)

Arguments

dat

An object of class osmdata

id

OSM identification of one or more objects for which lines are to be extracted

Value

An sf Simple Features Collection of linestrings

Examples

## Not run: ------------------------------------
# dat <- opq ("hengelo nl") %>% add_feature (key="highway") %>% osmdata_sf ()
# bus <- dat$osm_points [which (dat$osm_points$highway == 'bus_stop'),] %>%
#         rownames () # all OSM IDs of bus stops
# osm_lines (dat, bus) # all highways containing bus stops
# 
# # All lines which intersect with Piccadilly Circus in London, UK
# dat <- opq ("Fitzrovia London") %>% add_feature (key="highway") %>% 
#     osmdata_sf ()
# i <- which (dat$osm_polygons$name == "Piccadilly Circus")
# id <- rownames (dat$osm_polygons [i,])
# osm_lines (dat, id)
## ---------------------------------------------