Convert osmdata polygons into lines

Street networks downloaded with add_osm_object(key = "highway") will store any circular highways in $osm_polygons. this function combines those with the $osm_lines component to yield a single sf data.frame of all highways, whether polygonal or not.

osm_poly2line(osmdat)

Arguments

osmdat
An osmdata object.

Value

Modified version of same object with all $osm_polygons objeccts merged into $osm_lines.

Note

The $osm_polygons field is retained, with those features also repeated as LINESTRING objects in $osm_lines.

Examples

# NOT RUN {
dat <- opq ("colchester uk") %>%
            add_osm_feature (key="highway") %>%
            osmdata_sf ()
# colchester has lots of roundabouts, and these are stored in 'osm_polygons'
# rather than 'osm_lines'. The former can be merged with the latter by:
dat2 <- osm_poly2line (dat)
# 'dat2' will have more lines than 'dat', but the same number of polygons (they
# are left unchanged.) 
# }