Return an OSM Overpass query in XML format

Return an OSM Overpass query in XML format Read an (XML format) OSM Overpass response from a string, a connection, or a raw vector.

osmdata_xml(q, filename, quiet = TRUE, encoding)

Arguments

q

An object of class `overpass_query` constructed with opq and add_feature.

filename

If given, OSM data are saved to the named file

quiet

suppress status messages.

encoding

Unless otherwise specified XML documents are assumed to be encoded as UTF-8 or UTF-16. If the document is not UTF-8/16, and lacks an explicit encoding directive, this allows you to supply a default.

Value

An object of class `XML::xml_document` containing the result of the overpass API query.

Note

Objects of class xml_document can be saved as .xml or .osm files with codexml2::write_xml.