Issue OSM Overpass Query

Issue OSM Overpass Query

overpass_query(query, quiet = FALSE, wait = TRUE, pad_wait = 5,
  base_url = "http://overpass-api.de/api/interpreter", encoding)

Arguments

query

OSM Overpass query. Please note that the function is in ALPHA dev stage and needs YOU to specify that the output type is XML. However, you can use Overpass XML or Overpass QL formats.

quiet

suppress status messages. OSM Overpass queries may not return quickly. The package will display status messages by default showing when the query started/completed. You can disable these messages by setting this value to TRUE.

wait

if TRUE and if there is a queue at the Overpass API server, should this function wait and try again at the next available slot time or should it throw a an exception?

pad_wait

if there is a queue and wait is TRUE, pad the next query start time by pad_wait seconds (default = 5 seconds).

base_url

the url of the server running overpass to be queried, set to http://overpass-api.de/api/interpreter by default.

encoding

Unless otherwise specified XML documents are assumed to be encoded as UTF-8 or UTF-16. If the document is not UTF-8/16, and lacks an explicit encoding directive, this allows you to supply a default.

Value

If the query result only has OSM nodes then the function will return a SpatialPointsDataFrame with the nodes. If the query result has OSM ways then the function will return a SpatialLinesDataFrame with the ways relationss are not handled yet. If you asked for a CSV, you will receive the text response back, suitable for processing by read.table(text=..., sep=..., header=TRUE, check.names=FALSE, stringsAsFactors=FALSE).

Note

wrap function with httr::with_verbose if you want to see the httr query (useful for debugging connection issues).