unique_osmdata

Reduce the components of an osmdata object to only unique items of each type. That is, reduce $osm_points to only those points not present in other objects (lines, polygons, etc.); reduce $osm_lines to only those lines not present in multiline objects; and reduce $osm_polygons to only those polygons not present in multipolygon objects. This renders an osmdata object more directly compatible with typical output of sf.

unique_osmdata(dat)

Arguments

dat
An osmdata object

Value

Equivalent object reduced to only unique objects of each type