Add a feature to an Overpass query

Add a feature to an Overpass query

add_osm_feature(opq, key, value, key_exact = TRUE, value_exact = TRUE,
  match_case = TRUE, bbox = NULL)

add_feature(opq, key, value, key_exact = TRUE, value_exact = TRUE,
  match_case = TRUE, bbox = NULL)

Arguments

opq
An overpass_query object

key
feature key

value
value for feature key; can be negated with an initial exclamation mark, value = "!this", and can also be a vector, value = c ("this", "that").

key_exact
If FALSE, key is not interpreted exactly; see https://wiki.openstreetmap.org/wiki/Overpass_API

value_exact
If FALSE, value is not interpreted exactly

match_case
If FALSE, matching for both key and value is not sensitive to case

bbox
optional bounding box for the feature query; must be set if no opq query bbox has been set

Value

opq object

Note

key_exact should generally be TRUE, because OSM uses a reasonably well defined set of possible keys, as returned by available_features. Setting key_exact = FALSE allows matching of regular expressions on OSM keys, as described in Section 6.1.5 of http://wiki.openstreetmap.org/wiki/Overpass_API/Overpass_QL. The actual query submitted to the overpass API can be obtained from opq_string.

add_feature is deprecated; please use add_osm_feature.

References

http://wiki.openstreetmap.org/wiki/Map_Features

Examples

# NOT RUN {
q <- getbb ("portsmouth", display_name_contains = "United States") %>%
                opq () %>%
                add_osm_feature("amenity", "restaurant") %>%
                add_osm_feature("amenity", "pub")
osmdata_sf (q) # all objects that are restaurants AND pubs (there are none!)
q1 <- getbb ("portsmouth", display_name_contains = "United States") %>%
                opq () %>%
                add_osm_feature("amenity", "restaurant")
q2 <- getbb ("portsmouth", display_name_contains = "United States") %>%
                opq () %>%
                add_osm_feature("amenity", "pub")
c (osmdata_sf (q1), osmdata_sf (q2)) # all objects that are restaurants OR pubs
# Use of negation to extract all non-primary highways
q <- opq ("portsmouth uk") %>%
        add_osm_feature (key="highway", value = "!primary")
# }