osmdata class def

osmdata class def

The Overpass API (or OSM3S) is a read-only API that serves up custom selected parts of the OSM map data. It acts as a database over the web: the client sends a query to the API and gets back the data set that corresponds to the query.

Unlike the main API, which is optimized for editing, Overpass API is optimized for data consumers that need a few elements within a glimpse or up to roughly 100 million elements in some minutes, both selected by search criteria like e.g. location, type of objects, tag properties, proximity, or combinations of them. It acts as a database backend for various services.

Overpass API has a powerful query language (language guide, language reference, an IDE) beyond XAPI, but also has a compatibility layer to allow a smooth transition from XAPI.

osmdata(bbox, overpass_call, osm_points, osm_linestrings, osm_polygons,
  osm_multilinestrings, osm_multipolygons, timestamp, ...)

Arguments

bbox

bounding box

overpass_call

overpass_call

osm_points

sf Simple Features Collection of points

osm_linestrings

sf Simple Features Collection of multilinestrings

osm_polygons

sf Simple Features Collection of polygons

osm_multilinestrings

sf Simple Features Collection of multilinestrings

osm_multipolygons

sf Simple Features Collection of multipolygons

timestamp

timestamp

...

other options ignored

Note

Class constructor should never be used directly, and is only exported to provide access to the print method

Functions

add_feature

Add feature to an overpass API query

available_features

Obtain list all OpenStreetMap features

available_tags

List all tags for a given OpenStreetMap feature

bbox_to_string

Convert matrix or vector to a bbox string to pass to overpass API

issue_query

Finalize and issue an Overpass query

opq

Begin building an Overpass query

overpass_query

Issue OSM Overpass Query

overpass_status

Retrieve status of the Overpass API

read_osm

Read an XML OSM Overpass response from path